IT fim: added constants for diff_path 	moved the function generate_string to string.py 	uses WINDOWS constant
