+++ src/wazuh_testing/modules/fim/configuration.py
@@ -6,6 +6,7 @@
 SYSCHECK_DEBUG = 'syscheck.debug'
 SYMLINK_SCAN_INTERVAL = 'syscheck.symlink_scan_interval'
 FILE_MAX_SIZE = 'syscheck.file_max_size'
+LAST_ENTRY_FILE = 'last-entry.gz'
 
 # Attributes constants.
 ATTR_SHA1SUM = 'hash_sha1'
