+++ src/wazuh_testing/utils/file.py
@@ -594,32 +594,6 @@ def exists(path:str) -> bool:
     return os.path.exists(path) or os.path.islink(path)
 
 
-def generate_string(stringLength=10, character='0'):
-    """Generate a string with line breaks.
-
-    Parameters
-    ----------
-    stringLength : int, optional
-        Number of characters to add in the string. Default `10`
-    character : str, optional
-        Character to be added. Default `'0'`
-
-    Returns
-    -------
-    random_str : str
-        String with line breaks.
-    """
-    generated_string = ''
-
-    for i in range(stringLength):
-        generated_string += character
-
-        if i % 127 == 0:
-            generated_string += '\n'
-
-    return generated_string
-
-
 def translate_size(configured_size='1KB'):
     """
     Translate the configured size from string to number in bytes.
